add pretrained fastervit to pip package
