+++ README.md
@@ -29,45 +29,73 @@ cross-window carrier tokens.
 
 ![teaser](./fastervit/assets/hierarchial_attn.png)
 
-## 💥 News 💥
-
-- **[06.30.2023]** 🔥🔥 We have further improved the [TensorRT](https://developer.nvidia.com/tensorrt-getting-started) throughput of FasterViT models by 10-15% on average across different models.  
-- **[06.29.2023]** 🔥 Any-resolution FasterViT model can now be intitialized from pre-trained ImageNet resolution (224 x 244) models.
-- **[06.18.2023]** We have released the FasterViT [pip package](https://pypi.org/project/fastervit/) !
-- **[06.17.2023]** 🔥 [Any-resolution FasterViT](./fastervit/models/faster_vit_any_res.py)  model is now available ! the model can be used for variety of applications such as detection and segmentation or high-resolution fine-tuning with arbitrary input image resolutions.
-- **[06.09.2023]** 🔥🔥 We have released source code and ImageNet-1K FasterViT-models !
-
 ## Quick Start
 
-The FasterViT can be conveniently installed by:
+We can import pre-trained FasterViT models with **1 line of code**. First, FasterViT can be simply installed by:
 
 ```bash
 pip install fastervit
 ```
 
-A FasterViT model with default hyper-parameters can be created as in the following:
+A pretrained FasterViT model with default hyper-parameters can be created as in the following:
 
 ```python
 >>> from fastervit import create_model
 
 # Define fastervit-0 model with 224 x 224 resolution
->>> model = create_model('faster_vit_0_224')
+
+>>> model = create_model('faster_vit_0_any_res', 
+                          pretrained=True,
+                          model_path="/tmp/faster_vit_1.pth.tar")
+
+>>> model = create_model('faster_vit_0_224', pretrained=True)
+```
+We can adjust where to download the above FasterViT-0 model by passing `model_path`.
+
+We can also simply test the model by passing a dummy input image. The output is the logits:
+
+```python
+>>> import torch
+
+>>> image = torch.rand(1, 3, 224, 224)
+>>> output = model(image) # torch.Size([1, 1000])
 ```
 
-We can also use the any-resolution FasterViT model to accommodate arbitrary image resolutions. In the following, we define an any-resolution FasterViT-1
+We can also use the any-resolution FasterViT model to accommodate arbitrary image resolutions. In the following, we define an any-resolution FasterViT-0
 model with input resolution of 576 x 960, window sizes of 12 and 6 in 3rd and 4th stages, carrier token size of 2 and embedding dimension of
 128:
 
 ```python
 >>> from fastervit import create_model
 
-# Define any-resolution FasterViT-1 model with 576 x 960 resolution
->>> model = create_model('faster_vit_1_any_res', 
+# Define any-resolution FasterViT-0 model with 576 x 960 resolution
+>>> model = create_model('faster_vit_0_any_res', 
                           resolution=[576, 960],
                           window_size=[7, 7, 12, 6],
                           ct_size=2,
-                          dim=128)
+                          dim=64,
+                          pretrained=True)
 ```
+Note that the above model is intiliazed from the original ImageNet pre-trained FasterViT with original resolution of 224 x 224. As a result, missing keys and mis-matches could be expected since we are addign new layers (e.g. addition of new carrier tokens, etc.) 
+
+We can simply test the model by passing a dummy input image. The output is the logits:
+
+```python
+>>> import torch
+
+>>> image = torch.rand(1, 3, 576, 960)
+>>> output = model(image) # torch.Size([1, 1000])
+```
+
+## 💥 News 💥
+
+- **[07.04.2023]** 🔥🔥 ImageNet pretrained FasterViT models can now be imported with ** 1 line of code **. Please install the latest FasterViT pip package to use this functionality (also supports Any-resolution FasterViT models).
+- **[06.30.2023]** 🔥 We have further improved the [TensorRT](https://developer.nvidia.com/tensorrt-getting-started) throughput of FasterViT models by 10-15% on average across different models.  
+- **[06.29.2023]** Any-resolution FasterViT model can now be intitialized from pre-trained ImageNet resolution (224 x 244) models.
+- **[06.18.2023]** We have released the FasterViT [pip package](https://pypi.org/project/fastervit/) !
+- **[06.17.2023]** 🔥 [Any-resolution FasterViT](./fastervit/models/faster_vit_any_res.py)  model is now available ! the model can be used for variety of applications such as detection and segmentation or high-resolution fine-tuning with arbitrary input image resolutions.
+- **[06.09.2023]** 🔥🔥 We have released source code and ImageNet-1K FasterViT-models !
+
 
 ## Catalog
 - [x] ImageNet-1K training code
